feat: :sparkles: Adds a suite of functions to query the ABS API
`read_abs_api()` is the main function to query the API, and it is supported by a cast of helpers. `read_abs_api()` queries a URL's data and metadata, joins the two results, and returns a tidy `data.frame`.  The helper functions support infering the metadata URL from a query URL, tidying the returned data, checking a URL is valid, and breaking an invalid URL into multiple (hopefully) valid URLs.